chanstate: move htlc serialization
Move the HTLC serialization helpers into chanstate now that the shared
element codec is available there. Keep channeldb wrappers for the
exported helpers so existing callers continue to compile while follow-up
commits move more channel state code.
